move reference objects under security
