--- errata/templates/errata/erratumreference_list.html
@@ -1,7 +1,0 @@
-{% extends "objectlist.html" %}
-
-{% block page_title %}Erratum References{% endblock %}
-
-{% block breadcrumbs %} {{ block.super }} <li><a href="{% url 'errata:erratum_list' %}">Errata</a></li><li class="active"><a href="{% url 'errata:erratumreference_list' %}">Erratum References</a></li>{% endblock %}
-
-{% block content_title %} Erratum References {% endblock %}
