+++ errata/utils.py
@@ -58,65 +58,6 @@ def get_or_create_erratum(name, e_type, issue_date, synopsis):
     return e, created
 
 
-def fixup_erratum_reference(eref):
-    """ Fix up an ErratumReference object to normalize the URL and type
-    """
-    url = urlparse(eref.get('url'))
-    er_type = eref.get('er_type')
-    if 'lists' in url.hostname or 'lists' in url.path:
-        er_type = 'Mailing List'
-    if er_type == 'bugzilla' or 'bug' in url.hostname or 'bugs' in url.path:
-        er_type = 'Bug Tracker'
-    if ('ubuntu.com' in url.hostname and 'usn/' in url.path) or url.hostname == 'usn.ubuntu.com':
-        netloc = url.netloc.replace('usn.', '').replace('www.', '')
-        path = url.path.replace('usn/', 'security/notices/').replace('usn', 'USN').rstrip('/')
-        usn_id = path.split('/')[-1]
-        if 'USN' not in usn_id:
-            path = '/'.join(path.split('/')[:-1]) + '/USN-' + usn_id
-        url = url._replace(netloc=netloc, path=path)
-    if url.hostname == 'ubuntu.com' and url.path.startswith('/security/notices/USN'):
-        er_type = 'USN'
-    if 'launchpad.net' in url.hostname:
-        er_type = 'Bug Tracker'
-        netloc = url.netloc.replace('bugs.', '')
-        bug = url.path.split('/')[-1]
-        path = f'/bugs/{bug}'
-        url = url._replace(netloc=netloc, path=path)
-    if url.hostname in ['bugzilla.redhat.com', 'bugzilla.opensuse.org', 'bugs.suse.com'] and \
-            url.path == '/show_bug.cgi':
-        bug = url.query.split('=')[1]
-        path = f'/{bug}'
-        url = url._replace(path=path, query='')
-    if url.hostname == 'rhn.redhat.com':
-        netloc = url.netloc.replace('rhn', 'access')
-        path = url.path.replace('.html', '')
-        url = url._replace(netloc=netloc, path=path)
-    if url.hostname == 'access.redhat.com':
-        if 'l1d-cache-eviction-and-vector-register-sampling' in url.path or \
-                'security/vulnerabilities/speculativeexecution' in url.path or \
-                'security/vulnerabilities/stackguard' in url.path:
-            er_type = 'Link'
-        elif 'security/cve' in url.path:
-            return
-        else:
-            old_ref = url.path.split('/')[-1]
-            refs = old_ref.split('-')
-            if ':' not in url.path:
-                try:
-                    new_ref = f'{refs[0]}-{refs[1]}:{refs[2]}'
-                    path = url.path.replace(old_ref, new_ref)
-                    url = url._replace(path=path)
-                except IndexError:
-                    pass
-            er_type = refs[0].upper()
-    final_url = url.geturl()
-    if final_url in ['https://launchpad.net/bugs/', 'https://launchpad.net/bugs/XXXXXX']:
-        return
-    eref['er_type'] = er_type
-    eref['url'] = final_url
-    return eref
-
-
 def mark_errata_security_updates():
     """ For each set of erratum packages, modify any PackageUpdate that
         should be marked as a security update.
