+++ patchman/urls.py
@@ -44,8 +44,8 @@ router.register(r'package-name', package_views.PackageNameViewSet)
 router.register(r'package', package_views.PackageViewSet)
 router.register(r'package-update', package_views.PackageUpdateViewSet)
 router.register(r'cve', security_views.CVEViewSet)
+router.register(r'reference', security_views.ReferenceViewSet),
 router.register(r'erratum', errata_views.ErratumViewSet)
-router.register(r'erratum-reference', errata_views.ErratumReferenceViewSet)
 router.register(r'repo', repo_views.RepositoryViewSet)
 router.register(r'mirror', repo_views.MirrorViewSet)
 router.register(r'mirror-package', repo_views.MirrorPackageViewSet)
