+++ security/admin.py
@@ -15,9 +15,10 @@
 # along with Patchman. If not, see <http://www.gnu.org/licenses/>
 
 from django.contrib import admin
-from security.models import CWE, CVSS, CVE
+from security.models import CWE, CVSS, CVE, Reference
 
 
 admin.site.register(CWE)
 admin.site.register(CVSS)
 admin.site.register(CVE)
+admin.site.register(Reference)
