+++ security/templates/security/reference_list.html
@@ -1,0 +1,7 @@
+{% extends "objectlist.html" %}
+
+{% block page_title %}References{% endblock %}
+
+{% block breadcrumbs %} {{ block.super }} <li><a href="{% url 'security:security_landing' %}">Security</a></li><li class="active"><a href="{% url 'security:reference_list' %}">References</a></li>{% endblock %}
+
+{% block content_title %} References {% endblock %}
