+++ security/templates/security/security_landing.html
@@ -13,6 +13,7 @@
     <tr><td><a href="{% url 'security:cve_list' %}">CVEs</a></td></tr>
     <tr><td><a href="{% url 'security:cwe_list' %}">CWEs</a></td></tr>
     <tr><td><a href="{% url 'errata:erratum_list' %}?e_type=security">Security Errata</a></td></tr>
+    <tr><td><a href="{% url 'security:reference_list' %}">Security References</a></td></tr>
   </table>
 </div>
